+++ xdef/changelog.md
@@ -1,10 +1,13 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 42.1.6, release-date 2024-05-24
+# Version 42.1.6, release-date 2024-05-25
 * Fixed bug that did not report XDEF385 error when type name in local
   declaration section was the same as the referenced type.
 * Fixed bug generated negative value to `maxOccurs` attribute in 
   X-definition conversion to XML Schema.
+* If the name of root model in generator of XML schema from X-definition
+  (`org.xdef.util.XdefToXsd`) is set to `?type`, the only the schema with
+  declared simple types.
 
 # Version 42.1.5, release-date 2024-05-20
 * If the type name in the declaration section is the same as the referenced type